requirements.txt
django
gunicorn
python-dotenv
whitenoise
django-admin-datta==1.0.10
boto3
discord
requests
# psycopg2-binary
# mysqlclient

# Tooling
# django-dynamic-datatb==1.0.25
# django-api-generator==1.0.17
